Fix output type from classify cli

+++ src/surfclass/scripts/classify.py
@@ -67,10 +67,10 @@ def testmodel1(
     class_prediction = classifier.classify(X)
     logger.debug("Finished classification")
 
-    output = np.zeros(mask.shape[0])
+    output = np.zeros(mask.shape[0], dtype="uint8")
 
     # Convert to byte array to save space
-    output[indices] = np.int8(class_prediction)
+    output[indices] = np.uint8(class_prediction)
     output = output.reshape(_shape)
     name = f"{fileprefix}{filename}{filepostfix}.tif"
     outpath = str(pathlib.Path(outdir) / name)
@@ -289,10 +289,10 @@ def genericmodel(rasterfiles, model, outdir, bbox, prefix, postfix):
     class_prediction = classifier.classify(X)
     logger.debug("Finished classification")
 
-    output = np.zeros(mask.shape[0])
+    output = np.zeros(mask.shape[0], dtype="uint8")
 
     # Convert to byte array to save space
-    output[indices] = np.int8(class_prediction)
+    output[indices] = class_prediction.astype("uint8")
     output = output.reshape(_shape)
     name = f"{fileprefix}{filename}{filepostfix}.tif"
     outpath = str(pathlib.Path(outdir) / name)
